requirements.txt
python-dotenv
telethon